+++ lib/core/facets/array/span.rb
@@ -14,6 +14,11 @@ class Array
   # Returns a 2-tuple of arrays, the first element being the longest prefix for
   # which the block evaluates to true, the second element being the rest of the
   # array.
+  #
+  # TODO: This is based on Haskell's `span`, but the name is not intuitive
+  # for Ruby. Consider a more descriptive name (e.g. `split_while`).
+  # Also consider adding the inverse (Haskell's `break`), which splits
+  # where the predicate first becomes true.
   def span
     # This implementation is another way of saying
     # `[take_while(&:block), drop_while(&:block)]`