+++ lib/core/facets/array/step.rb
@@ -2,30 +2,28 @@ class Array
 
   # Iterate over every nth element of an array.
   #
+  # An optional offset can be given to start from a
+  # specific index (default is 0).
+  #
+  #   [:a, :b, :c, :d, :e, :f].step(2).to_a    #=> [:a, :c, :e]
+  #   [:a, :b, :c, :d, :e, :f].step(2, 1).to_a  #=> [:b, :d, :f]
+  #   [:a, :b, :c, :d, :e, :f].step(3).to_a    #=> [:a, :d]
+  #
   #   r = []
   #   [:a, :b, :c, :d].step(2) { |x| r << x }
-  #   r   #=> [:b, :d]
-  #   
-  # Without a block it returns an Enumerator.
-  #
-  #   [:a, :b, :c, :d].step(1).to_a   #=> [:a, :b, :c, :d]
-  #   [:a, :b, :c, :d].step(2).to_a   #=> [:b, :d]
-  #   [:a, :b, :c, :d].step(3).to_a   #=> [:c]
-  #   [:a, :b, :c, :d].step(5).to_a   #=> []
+  #   r   #=> [:a, :c]
   #
   # CREDIT: Ryan Duryea
 
-  def step(n) #:yield:
+  def step(n, offset=0) #:yield:
     if block_given?
-      ((n - 1)...size).step(n).each do |i|
+      (offset...size).step(n).each do |i|
         yield(fetch(i))
       end
     else
-      Enumerator.new(size / n) do |y|
-        ((n - 1)...self.size).step(n).each { |i| y << fetch(i) }  
+      Enumerator.new do |y|
+        (offset...size).step(n).each { |i| y << fetch(i) }
       end
     end
   end
 end
-
-